# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Admin:
    properties:
      accountId:
        type: integer
      fio:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
    type: object
  entities.Group:
    properties:
      id:
        type: integer
      name:
        type: string
      teacherId:
        type: integer
    type: object
...
  requests.CreateAdminRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      phone_number:
        type: string
    type: object
  requests.CreateGroupRequest:
    properties:
      name:
        type: string
      teacher_id:
        type: integer
    type: object
...
  requests.UpdateAdminRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      id:
        type: integer
      phone_number:
        type: string
    type: object
  requests.UpdateGroupRequest:
    properties:
      id:
        type: integer
      name:
        type: string
      teacher_id:
        type: integer
    type: object
...
  usecases.ReadAdminResponseDto:
    properties:
      admin:
        $ref: '#/definitions/entities.Admin'
    type: object
  usecases.ReadAllGroupsResponseDto:
    properties:
      groups:
        items:
          $ref: '#/definitions/entities.Group'
        type: array
    type: object
...
  usecases.ReadGroupResponseDto:
    properties:
      group:
        $ref: '#/definitions/entities.Group'
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: Backend for KeenEye
  title: Backend for KeenEye
  version: 1.0.0
paths:
...
  /api/create-admin:
    post:
      consumes:
      - application/json
      description: Create a new admin (admin only)
      parameters:
      - description: Admin info
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/requests.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Admin'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Create admin
      tags:
      - admins
  /api/create-group:
    post:
      consumes:
      - application/json
      description: Create a new group (admin only)
      parameters:
      - description: Group info
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/requests.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Group'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Create group
      tags:
      - groups
...
  /api/delete-admin:
    delete:
      description: Delete admin by ID (admin only)
      parameters:
      - description: Admin ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid admin ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Delete admin
      tags:
      - admins
  /api/delete-group:
    delete:
      description: Delete group by ID (admin only)
      parameters:
      - description: Group ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid group ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Delete group
      tags:
      - groups
...
  /api/read-admin:
    get:
      description: Get admin by ID (admin only)
      parameters:
      - description: Admin ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadAdminResponseDto'
        "400":
          description: Invalid admin ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get admin by ID
      tags:
      - admins
  /api/read-all-groups:
    get:
      description: Get list of all groups (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadAllGroupsResponseDto'
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get all groups
      tags:
      - groups
...
  /api/read-group:
    get:
      description: Get group by ID (student, teacher or admin). Students and teachers
        are allowed to access only their group.
      parameters:
      - description: Group ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadGroupResponseDto'
        "400":
          description: Invalid group ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get group by ID
      tags:
      - groups
...
  /api/update-admin:
    put:
      consumes:
      - application/json
      description: Update admin info (admin only)
      parameters:
      - description: Updated admin info
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Admin'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Update admin
      tags:
      - admins
  /api/update-group:
    put:
      consumes:
      - application/json
      description: Update group info (admin only)
      parameters:
      - description: Updated group info
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Group'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Update group
      tags:
      - groups
...
securityDefinitions:
  BasicAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"